+++ examples/hubbard_relative_overlap.cpp
@@ -10,7 +10,7 @@
 
 struct CliOptions {
   size_t lattice_size = 16;
-  size_t total_momentum = 0;
+  int64_t total_momentum = 0;
   double t = 1.0;
   double U = -4.0;
   size_t num_steps = 100;
@@ -22,7 +22,7 @@ void parse_cli_options(int argc, char** argv, CliOptions* options_out) {
   // clang-format off
   options.add_options()
       ("L,lattice-size", "Lattice size per dimension",  cxxopts::value<size_t>()->default_value("16"))
-      ("P,total-momentum", "Total momentum value",      cxxopts::value<size_t>()->default_value("0"))
+      ("P,total-momentum", "Total momentum value",      cxxopts::value<int64_t>()->default_value("0"))
       ("t,hopping", "Hopping amplitude",                cxxopts::value<double>()->default_value("1.0"))
       ("U,interaction", "On-site interaction strength", cxxopts::value<double>()->default_value("-4.0"))
       ("n,num-steps", "Number of normalization steps",  cxxopts::value<size_t>()->default_value("100"))
@@ -36,7 +36,7 @@ void parse_cli_options(int argc, char** argv, CliOptions* options_out) {
       std::exit(0);
     }
     options_out->lattice_size = result["lattice-size"].as<size_t>();
-    options_out->total_momentum = result["total-momentum"].as<size_t>();
+    options_out->total_momentum = result["total-momentum"].as<int64_t>();
     options_out->t = result["hopping"].as<double>();
     options_out->U = result["interaction"].as<double>();
     options_out->num_steps = result["num-steps"].as<size_t>();
@@ -52,8 +52,8 @@ int main(int argc, char** argv) {
   parse_cli_options(argc, argv, &opts);
 
   const std::vector<size_t> lattice_size{opts.lattice_size, opts.lattice_size, opts.lattice_size};
-  const std::vector<size_t> total_momentum{opts.total_momentum, opts.total_momentum,
-                                           opts.total_momentum};
+  const std::vector<int64_t> total_momentum{opts.total_momentum, opts.total_momentum,
+                                            opts.total_momentum};
 
   HubbardRelative hamiltonian(lattice_size, total_momentum, opts.t, opts.U);
